feat: 기본적인 패킷 클래스 구현

+++ SpikeBeach/Source/SpikeBeach/MultiplaySystem/WebSocketPackets/Notifies/GameStartNotify.h
@@ -13,7 +13,7 @@ class SPIKEBEACH_API GameStartNotify : public NotifyHeader
 public:
     FString gameInfoString;
     FString ipAddress;
-    int portNumber;
+    uint16 portNumber;
 
     TArray<uint8> Serialize();
     virtual int Deserialize(const uint8* data);
